requirements.txt
chatterbox-tts
soundfile
librosa
huggingface_hub
perth
